requirements/requirements_dev.txt
-r base.txt

black
flake8
django-debug-toolbar
isort
mypy
pytest

